+++ app/routes/app.switcher/route.tsx
@@ -29,6 +29,8 @@ import { isShopMigrated } from "~/server/translateV4/migration.server";
 import {
   loadSwitcherConfigCompat,
   saveSwitcherConfigCompat,
+  buildSwitcherEditDefaults,
+  type SwitcherEditData,
 } from "./switcherClient";
 import { useSelector } from "react-redux";
 import { InfoCircleOutlined, SettingOutlined } from "@ant-design/icons";
@@ -41,21 +43,6 @@ import useReport from "scripts/eventReport";
 import CloseIcon from "~/components/icon/closeIcon";
 import { withEmbeddedSearch } from "~/utils/embeddedAction";
 import AppPageHeader from "~/ui/components/AppPageHeader";
-interface EditData {
-  shopName: string;
-  includedFlag: boolean;
-  languageSelector: boolean;
-  currencySelector: boolean;
-  ipOpen: boolean;
-  fontColor: string;
-  backgroundColor: string;
-  buttonColor: string;
-  buttonBackgroundColor: string;
-  optionBorderColor: string;
-  selectorPosition: string;
-  positionData: string;
-  isTransparent: boolean;
-}
 
 const initialLocalization = {
   languages: [
@@ -186,8 +173,8 @@ const Index = () => {
   const [isCurrencyOpen, setIsCurrencyOpen] = useState(false);
   const [isSelectorOpen, setIsSelectorOpen] = useState(false);
   const [localization, setLocalization] = useState(initialLocalization);
-  const [originalData, setOriginalData] = useState<EditData>();
-  const [editData, setEditData] = useState<EditData>({
+  const [originalData, setOriginalData] = useState<SwitcherEditData>();
+  const [editData, setEditData] = useState<SwitcherEditData>({
     shopName: "",
     includedFlag: false,
     languageSelector: false,
@@ -273,21 +260,7 @@ const Index = () => {
         shop,
         server: server as string,
       });
-      const initData = {
-        shopName: shop,
-        includedFlag: true,
-        languageSelector: true,
-        currencySelector: true,
-        ipOpen: false,
-        fontColor: "#303030",
-        backgroundColor: "#ffffff",
-        buttonColor: "#ffffff",
-        buttonBackgroundColor: "#f6f6f7",
-        optionBorderColor: "#d4d4d8",
-        selectorPosition: "bottom_left",
-        positionData: "10",
-        isTransparent: false,
-      };
+      const initData = buildSwitcherEditDefaults(shop);
       if (data?.success && data.response) {
         const filteredResponse = Object.fromEntries(
           Object.entries(data.response).filter(([_, value]) => value !== null),
@@ -360,28 +333,6 @@ const Index = () => {
       }
 
       setCardLoading(false);
-      // const switcherData =
-      //   themeFetcher.data.data.nodes[0].files.nodes[0].body.content;
-      // const jsonString = switcherData.replace(/\/\*[\s\S]*?\*\//g, "").trim();
-      // const themeData = JSON.parse(jsonString);
-
-      // const footer = themeData.sections?.footer;
-      // if (footer?.blocks) {
-      //   const switcherJson: any = Object.values(footer.blocks).find(
-      //     (block: any) => block.type === ciwiSwitcherBlocksId,
-      //   );
-
-      //   if (switcherJson) {
-      //     setSwitcherEnableCardOpen(true);
-      //   }
-      // }
-      // const isAppEnabled = Object.values(themeData.sections).some((section: any) =>
-      //   section?.blocks && Object.values(section.blocks).some((block: any) =>
-      //     block.type.includes(ciwiSwitcherBlocksId)
-      //   )
-      // );
-
-      // setSwitcherEnableCardOpen(isAppEnabled);
     }
   }, [themeFetcher.data]);
 
@@ -446,7 +397,6 @@ const Index = () => {
           }
         }
       }
-    } else {
     }
   }, [shopFetcher.data]);
 
@@ -462,7 +412,7 @@ const Index = () => {
     }
   }, [editData, originalData]);
 
-  const handleEditData = (updates: Partial<EditData>) => {
+  const handleEditData = (updates: Partial<SwitcherEditData>) => {
     // 更新对应的状态
     Object.entries(updates).forEach(([key, value]) => {
       switch (key) {
